Client/server sender split
